+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file. The format 
 
 <hr />
 
+## [2.1.22] - 2025-12-18
+Summary: Engine.
+
+### Changed
+- Touched areas: engine.
+
+### Files
+- includes/admin/schema-service.php
+- includes/admin/schema-view-settings.php
+- includes/admin/schema-view.php
+- includes/engine/core-elementor.php
+- includes/engine/core-helpers.php
+- includes/engine/core-site-entities.php
+
+
 ## [2.1.21] - 2025-12-18
 Summary: Engine.